+++ services/dumper.js
@@ -58,7 +58,7 @@ function Dumper(project, config) {
       'serve-favicon': '~2.3.0',
       'dotenv': '~2.0.0',
       'chalk': '~1.1.3',
-      'sequelize': '~3.24.8',
+      'sequelize': '3.30.4',
       'forest-express-sequelize': 'latest'
     };