+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [Unreleased]
 ### Added
 - MSSQL Support - Support Lumber generation on a MSSQL database hosted on Windows Azure.
+- MSSQL Support - Support new database types (MONEY, NTEXT, NVARCHAR).
+
+### Fixed
+- MSSQL Support - BIT types are now detected as Boolean instead of Number.